// samples/bookshop-web/src/types/fontSize.type.ts
export enum FontSize {
  twoXs = "twoXs", // 11
  xs = "xs", // 12
  sm = "sm", // 14
  md = "md", // 16
  lg = "lg", // 18
  xl = "xl", // 20
  twoXl = "twoXl", // 22
  threeXl = "threeXl", // 24
  fourXl = "fourXl", // 26
  fiveXl = "fiveXl", // 28
  sixXl = "sixXl", // 30
  sevenXl = "sevenXl", // 32
  eightXl = "eightXl", // 34
  nineXl = "nineXl", // 36
  tenXl = "tenXl", // 38
  elevenXl = "elevenXl", // 40
}